- made AnimSet's vel parameter an actual double.
This does not affect any outcome, it just avoids pointless type conversions.

+++ source/games/sw/src/game.h
@@ -1609,8 +1609,8 @@ void RefreshInfoLine(PLAYER* pp);
 void DoAnim(int numtics);
 void AnimDelete(int animtype, int animindex, DSWActor*);
 short AnimGetGoal(int animtype, int animindex, DSWActor*);
-int AnimSet(int animtype, int animindex, DSWActor* animactor, double thegoal, int thevel);
-int AnimSet(int animtype, sectortype* animindex, double thegoal, int thevel)
+int AnimSet(int animtype, int animindex, DSWActor* animactor, double thegoal, double thevel);
+int AnimSet(int animtype, sectortype* animindex, double thegoal, double thevel)
 {
     return AnimSet(animtype, sectnum(animindex), nullptr, thegoal, thevel);
 }
